# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3
 * Do not edit manually.
 * Api
 * SkillForge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  bio?: string | null;
  avatarUrl?: string | null;
  githubUrl?: string | null;
  skills: string[];
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  looking_for_team: "looking_for_team",
  active: "active",
  completed: "completed",
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  techStack: string[];
  teamSize: number;
  status: ProjectStatus;
  githubUrl?: string | null;
  ownerId: number;
  ownerName: string;
  ownerAvatarUrl?: string | null;
  memberCount: number;
  isBookmarked: boolean;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  bio?: string | null;
  avatarUrl?: string | null;
  githubUrl?: string | null;
  skills: string[];
  createdAt: string;
  projects: Project[];
}

export interface UsersListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface UpdateProfileRequest {
  name?: string;
  bio?: string;
  avatarUrl?: string;
  githubUrl?: string;
  skills?: string[];
}

export type ProjectDetailStatus =
  (typeof ProjectDetailStatus)[keyof typeof ProjectDetailStatus];

export const ProjectDetailStatus = {
  looking_for_team: "looking_for_team",
  active: "active",
  completed: "completed",
} as const;

export interface ProjectDetail {
  id: number;
  title: string;
  description: string;
  techStack: string[];
  teamSize: number;
  status: ProjectDetailStatus;
  githubUrl?: string | null;
  ownerId: number;
  ownerName: string;
  ownerAvatarUrl?: string | null;
  memberCount: number;
  isBookmarked: boolean;
  createdAt: string;
  members: User[];
}

export interface ProjectsListResponse {
  projects: Project[];
  total: number;
  page: number;
  limit: number;
}

export type CreateProjectRequestStatus =
  (typeof CreateProjectRequestStatus)[keyof typeof CreateProjectRequestStatus];

export const CreateProjectRequestStatus = {
  looking_for_team: "looking_for_team",
  active: "active",
  completed: "completed",
} as const;

export interface CreateProjectRequest {
  title: string;
  description: string;
  techStack: string[];
  teamSize: number;
  status: CreateProjectRequestStatus;
  githubUrl?: string;
}

export type UpdateProjectRequestStatus =
  (typeof UpdateProjectRequestStatus)[keyof typeof UpdateProjectRequestStatus];

export const UpdateProjectRequestStatus = {
  looking_for_team: "looking_for_team",
  active: "active",
  completed: "completed",
} as const;

export interface UpdateProjectRequest {
  title?: string;
  description?: string;
  techStack?: string[];
  teamSize?: number;
  status?: UpdateProjectRequestStatus;
  githubUrl?: string;
}

export interface CreateCollaborationRequest {
  projectId: number;
  message?: string;
}

export type RespondToRequestBodyAction =
  (typeof RespondToRequestBodyAction)[keyof typeof RespondToRequestBodyAction];

export const RespondToRequestBodyAction = {
  accept: "accept",
  reject: "reject",
} as const;

export interface RespondToRequestBody {
  action: RespondToRequestBodyAction;
}

export type CollaborationRequestStatus =
  (typeof CollaborationRequestStatus)[keyof typeof CollaborationRequestStatus];

export const CollaborationRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface CollaborationRequest {
  id: number;
  projectId: number;
  requesterId: number;
  message?: string | null;
  status: CollaborationRequestStatus;
  createdAt: string;
}

export type CollaborationRequestDetailStatus =
  (typeof CollaborationRequestDetailStatus)[keyof typeof CollaborationRequestDetailStatus];

export const CollaborationRequestDetailStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface CollaborationRequestDetail {
  id: number;
  projectId: number;
  projectTitle: string;
  requesterId: number;
  requesterName: string;
  requesterAvatarUrl?: string | null;
  requesterSkills: string[];
  message?: string | null;
  status: CollaborationRequestDetailStatus;
  createdAt: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  collaboration_request: "collaboration_request",
  request_accepted: "request_accepted",
  request_rejected: "request_rejected",
  project_invite: "project_invite",
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  message: string;
  relatedProjectId?: number | null;
  relatedProjectTitle?: string | null;
  isRead: boolean;
  createdAt: string;
}

export type GetUsersParams = {
  search?: string;
  skill?: string;
  page?: number;
  limit?: number;
};

export type GetProjectsParams = {
  search?: string;
  tech?: string;
  status?: GetProjectsStatus;
  page?: number;
  limit?: number;
};

export type GetProjectsStatus =
  (typeof GetProjectsStatus)[keyof typeof GetProjectsStatus];

export const GetProjectsStatus = {
  looking_for_team: "looking_for_team",
  active: "active",
  completed: "completed",
} as const;