# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /collaborations/request:
    post:
      operationId: createCollaborationRequest
      tags: [collaborations]
      summary: Request to join a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollaborationRequest"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationRequest"
        "400":
          description: Already requested or member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collaborations/requests/incoming:
    get:
      operationId: getIncomingRequests
      tags: [collaborations]
      summary: Get collaboration requests for owner's projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Incoming requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollaborationRequestDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collaborations/requests/outgoing:
    get:
      operationId: getOutgoingRequests
      tags: [collaborations]
      summary: Get collaboration requests sent by current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Outgoing requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollaborationRequestDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collaborations/requests/{requestId}/respond:
    put:
      operationId: respondToCollaborationRequest
      tags: [collaborations]
      summary: Accept or reject a collaboration request
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToRequestBody"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: Get current user's bookmarked projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bookmarked projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookmarks/{projectId}:
    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Bookmark a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Bookmark removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - skills
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        techStack:
          type: array
          items:
            type: string
        teamSize:
          type: number
        status:
          type: string
          enum: [looking_for_team, active, completed]
        githubUrl:
          type: string
          nullable: true
        ownerId:
          type: number
        ownerName:
          type: string
        ownerAvatarUrl:
          type: string
          nullable: true
        memberCount:
          type: number
        isBookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - techStack
        - teamSize
        - status
        - ownerId
        - ownerName
        - memberCount
        - isBookmarked
        - createdAt
...
    CreateCollaborationRequest:
      type: object
      properties:
        projectId:
          type: number
        message:
          type: string
      required:
        - projectId
    RespondToRequestBody:
      type: object
      properties:
        action:
          type: string
          enum: [accept, reject]
      required:
        - action
    CollaborationRequest:
      type: object
      properties:
        id:
          type: number
        projectId:
          type: number
        requesterId:
          type: number
        message:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - requesterId
        - status
        - createdAt
    CollaborationRequestDetail:
      type: object
      properties:
        id:
          type: number
        projectId:
          type: number
        projectTitle:
          type: string
        requesterId:
          type: number
        requesterName:
          type: string
        requesterAvatarUrl:
          type: string
          nullable: true
        requesterSkills:
          type: array
          items:
            type: string
        message:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - projectTitle
        - requesterId
        - requesterName
        - requesterSkills
        - status
        - createdAt
...