# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SkillForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  bio?: string | null;
  avatarUrl?: string | null;
  githubUrl?: string | null;
  skills: string[];
  createdAt: Date;
}